int32 -> int64

+++ SARMS/Users/Administrator.cs
@@ -611,7 +611,7 @@ namespace SARMS.Users
 
                     if (reader.HasRows)
                     {
-                        Unit temp = new Unit(Convert.ToInt32(reader[0]), reader[1].ToString(), reader[2].ToString(), Convert.ToInt16(reader[3]), Convert.ToByte(reader[4]), Convert.ToInt32(reader[5]), Convert.ToInt32(reader[6]));
+                        Unit temp = new Unit(Convert.ToInt64(reader[0]), reader[1].ToString(), reader[2].ToString(), Convert.ToInt16(reader[3]), Convert.ToByte(reader[4]), Convert.ToInt32(reader[5]), Convert.ToInt32(reader[6]));
                         temp.Assessments = GetUnitAssessments(temp);
                         return temp;
                     }
